Correcting the system lists across the site. All systems should be displayed like this:
(NICKNAME) Full System Name

And should be sorted by NICKNAME. To accomplish this I overloaded
fisma_model's base method getList() in the system model to format
the system list correctly.

I didn't get to test this too thoroughly, but the user create,
edit, and view pages should be working, as are the system lists
in finding, remediation search, and a few of the reports.

I'll do more thorough testing tomorrow.