bug fix on branch #2776282 - Hide Decommissioned Systems
Made a good deal of modifications to the various reports for the new way the getList function works.

On some reports i had to add an if (isset()) for certain variables in the loops because the new returning value of getList
offset the index count. Hidden Systems do not show on any of these reports any longer