Feature request on the IR branch: OFJ-3 Incident Reporting Module
Found a bug on the clone incident feature. After you clone, if you refresh the page then it will continue to clone. Even if you go to another page and then click back, it clones the incident!!

This is indicative of a larger problem, which is that data is being modified in a GET request. But no time to fix that now, so for the time being I'm replacing the _forward() with a _redirect().