OFJ-303 Convert flash charts and graphs to a standard technology (PNG, SVG, or canvas)
Placed in new javascript lib for canvas based charts (jqPlot).
Created a javascript wrapper for jqPlot bar, stacked-bar, and pie charts.
Implemented a new php class called Fisma_ChartJQP (php wrapper) which hands data to the javascript wrapper.
Added json context switches to DashboardController.php (for use as a remote data source for the charts on the main page).
Changed the structure of outputed data from the totalstatusAction() and totaltypeAction() to match that of what the new charts expect.