OFJ-303 Convert flash charts and graphs to a standard technology (PNG, SVG, or canvas)
The Chart wrapper will now make a uniqueid for itself if one is not supplied to it (no longer a required argument).
Added alignment option to the chart wrapper.
Changed the data output structure of http://localhost/security-control-chart/control-deficiencies/format/json
Replaced the chart on the Findings dashboard with the new chart implementation