OFJ-963 Executive dashboard
Extended the chart wrapper further to support "widgets" (textboxes and dropdown combo-boxes) that will, once changed,
fade out the chart, reload the data from the remote data source (while pushing the data from the visible variables to
the controller), and represent the chart.

"widget" variables are also stored in cookies, so when a user selects a certain mode for a chart, it will retain that
setting for the next page load.

This is in preparation for the interactive exec-dashboard charts.