OFJ-963 Executive dashboard
Extended the chart JavaScript wrapper to auto-detect when the chart will need to be scrollable based on determining the width of each bar in a bar-graph for the space the chart is given. It will now create a scrollable div when nessesary (this is for the Findings Dashboard chart that displays the count of findings for each security controle).
Cleaned up the chart view-script code so its not painfull to read.