OFJ-963 Executive dashboard
Made the chart class more object oriented. The class no longer expects an array to be passed to it. Functions are used to set data within the class.
Added a addColumn() function to the class where data to be plotted and a link may be added at the same time.