OFJ-963 Executive dashboard
Added the Open Findings By Organization chart, with the ability to show counts of findings of high,mod,low threat levels for
agencies, bureaus, organizations, and systems. The finding counts are totaled with a given target/organizations
child organizations and their child organizations and so on...

Fixed a bug where widget settings colided with other widget settings of other charts where the widget names were the same

Apparently I enabeled a feature in my text editor today that automatically erases white space on lines that only have spaces,
hence the spam of \n-\n+ in the diff, I apologise for that. I have disabeled this feature, and it will not happen again.