OFJ-963 Executive dashboard
Added in jqPlotPlugin for axis-labels (a label for an entire axis, not for each individual columns and rows)
Added ability to controle the angle at which the x-axis column labels are tilted (set to 0 for no tilt)
Removed "days ago" from every column label that had something like "30-40 days ago", and untilted it, changed it to just "30-40", and added in a x-axis labels stating the measument is in days
Removed dates from column labels and changed them to day ranges (as requested last tuesday).
Added in y-axis labels stating "number of findings" for charts where it was nessesary (as requested last tuesday).
Removed needless decimals from left-side row-labels, row-labels will now be rounded.