OFJ-1226 Fix advanced search rendering in IE
Update ie7 hacks to fix most of the sizing/alignment problems in IE7. The alignment of the +/- buttons required a change to the javascript to add the elements in a different order.

Also made the default width of the advanced search stuff slightly wider, since I noticed with some lengthy criteria it will actually line wrap (which isn't good).