OFJ-1330 / OFC-346 Apache Authentication
As requested ( http://crucible.openfisma.org/cru/OFC-346#c2392 ), am making
root special in the sence that OpenFISMA will authenticate this account
against the database, rather than accept Apache's authentication.

With this change however, administrators will have to be carefull to make
sure Apache's password for root, and OpenFISMA's password for root is the
same. If they are ever different, root is broken - but because I forsee
this exact problem problem occuring in the future somewhere, I have changed
OpenFISMA's authentication UI to notice when the BasicAuth module is not on,
or if Apache somehow passed authentication but did not supply authentication
information to PHP; and show the regular log-in screen so root can be logged
into.

WARNING! This does not guarantee an OpenFISMA administrator can gain access
to root. If Apache's authentication fails, the entire domain is inaccessible
and the OpenFISMA scripts are not even executed.